+++ src/utils.rs
@@ -1,8 +1,8 @@
-pub(crate) fn padding_for_8_bytes(value: u32) -> u32 {
+pub(crate) fn padding_for_8_bytes(value: usize) -> usize {
     padding_for_n_bytes(value, 8)
 }
 
-pub(crate) fn padding_for_n_bytes(value: u32, align: u32) -> u32 {
+pub(crate) fn padding_for_n_bytes(value: usize, align: usize) -> usize {
     let len_rounded_up = value.wrapping_add(align).wrapping_sub(1) & !align.wrapping_sub(1);
 
     len_rounded_up.wrapping_sub(value)
